added delete authorization

+++ controllers/blogs.js
@@ -2,14 +2,7 @@ const jwt = require("jsonwebtoken");
 const blogRouter = require("express").Router();
 const Blog = require("../models/blog");
 const User = require("../models/user");
-
-const getTokenFrom = (require) => {
-	const authorization = router.get("authorization");
-	if (authorization && authorization.startsWith("Bearer ")) {
-		return authorization.replace("Bearer ", "");
-	}
-	return null;
-};
+const helper = require("../utils/middleware");
 
 blogRouter.get("/", async (request, response) => {
 	const blogs = await Blog.find({}).populate("user", { username: 1, name: 1 });
@@ -19,22 +12,22 @@ blogRouter.get("/", async (request, response) => {
 blogRouter.post("/", async (request, response, next) => {
 	const body = request.body;
 
-	const decodedToken = jwt.verify(getTokenFrom(request), process.env.SECRET);
+	const decodedToken = jwt.verify(
+		helper.getTokenFrom(request),
+		process.env.SECRET
+	);
 
 	if (!decodedToken.id) {
 		return response.status(401).json({ error: "token invalid" });
 	}
 
-	const user = await User.findbyId(decodedToken.id).populate("blogs");
+	const user = await User.findById(decodedToken.id).populate("blogs");
 
 	console.log("users", user);
-	console.log("first user", user[0]);
-
-	console.log("first users blogs", user[0].blogs);
 
 	const blog = new Blog({
 		title: body.title,
-		user: user[0]._id,
+		user: user._id,
 		url: body.url,
 		likes: body.likes,
 	});
@@ -42,8 +35,8 @@ blogRouter.post("/", async (request, response, next) => {
 	try {
 		const savedBlog = await blog.save();
 		console.log(savedBlog);
-		user[0].blogs = user[0].blogs.concat(savedBlog._id);
-		await user[0].save();
+		user.blogs = user.blogs.concat(savedBlog._id);
+		await user.save();
 		response.status(201).json(savedBlog);
 	} catch (exception) {
 		next(exception);
@@ -51,7 +44,30 @@ blogRouter.post("/", async (request, response, next) => {
 });
 
 blogRouter.delete("/:id", async (request, response) => {
-	await Blog.findByIdAndRemove(request.params.id);
+	console.log(request.params.id);
+	const decodedToken = jwt.verify(
+		helper.getTokenFrom(request),
+		process.env.SECRET
+	);
+
+	console.log(decodedToken);
+
+	const blog = await Blog.findById(request.params.id);
+
+	console.log(blog);
+
+	if (decodedToken.id === blog.user.toString()) {
+		const deletedBlog = await Blog.findByIdAndRemove(request.params.id);
+		response
+			.status(200)
+			.json({ message: "successfully deleted blog", deletedBlog });
+	} else {
+		response
+			.status(400)
+			.json({
+				error: "invalid attempt; logged in user does not have authorization",
+			});
+	}
 	response.status(204).end();
 });
